Add files via upload
This file works with groups of files from a URL. Takes each of the files and plots 2 graphs of irradiance vs time for diode A and Diode C (this is background work as the plots aren't shown). Then for each of those graphs, it finds the area under the curve, which is the integrated energy. All the final integrated energies are saved to an output file for ease of access.